requirements.txt
mediapipe==0.10.11
numpy==1.23.5
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.5.2
protobuf==3.20.3
fastapi==0.103.2
pillow==10.4.0
requests==2.32.3
uvicorn==0.32.0
python-multipart==0.0.16
imageio==2.36.0
pillow==10.4.0
